[BE] Service, ServiceImpl anti-pattern lecserélése csak az implementációra, Tranzakciók átnézése a Service-kben #24
