+++ src/main/java/hu/dominikvaradi/sociallybackend/flows/comment/domain/dto/CommentResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
 import java.time.ZonedDateTime;
-import java.util.Set;
+import java.util.List;
 import java.util.UUID;
 
 @Builder
@@ -35,5 +35,5 @@ public class CommentResponseDto {
 	private ZonedDateTime created;
 
 	@NotNull
-	private Set<ReactionCountResponseDto> reactionsCount;
+	private List<ReactionCountResponseDto> reactionsCount;
 }
