+++ src/main/java/hu/dominikvaradi/sociallybackend/flows/message/domain/dto/MessageResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
 import java.time.ZonedDateTime;
-import java.util.Set;
+import java.util.List;
 import java.util.UUID;
 
 @Builder
@@ -36,5 +36,5 @@ public class MessageResponseDto {
 	private ZonedDateTime created;
 
 	@NotNull
-	private Set<ReactionCountResponseDto> reactionsCount;
+	private List<ReactionCountResponseDto> reactionsCount;
 }
