+++ src/main/java/hu/dominikvaradi/sociallybackend/flows/post/domain/dto/PostResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
 import java.time.ZonedDateTime;
-import java.util.Set;
+import java.util.List;
 import java.util.UUID;
 
 @Builder
@@ -44,7 +44,7 @@ public class PostResponseDto {
 	private ZonedDateTime created;
 
 	@NotNull
-	private Set<ReactionCountResponseDto> reactionsCount;
+	private List<ReactionCountResponseDto> reactionsCount;
 
 	@NotNull
 	private long commentsCount;
