--- src/main/java/hu/dominikvaradi/sociallybackend/flows/security/service/JwtUserDetailsServiceImpl.java
@@ -1,23 +1,0 @@
-package hu.dominikvaradi.sociallybackend.flows.security.service;
-
-import hu.dominikvaradi.sociallybackend.flows.security.domain.JwtUserDetails;
-import hu.dominikvaradi.sociallybackend.flows.user.domain.User;
-import hu.dominikvaradi.sociallybackend.flows.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class JwtUserDetailsServiceImpl implements JwtUserDetailsService {
-	private final UserRepository userRepository;
-
-	@Override
-	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-		User user = userRepository.findByEmail(email)
-				.orElseThrow(() -> new UsernameNotFoundException("USER_NOT_FOUND"));
-
-		return new JwtUserDetails(user);
-	}
-}
