+++ src/main/java/hu/dominikvaradi/sociallybackend/flows/security/service/RefreshTokenService.java
@@ -1,19 +1,57 @@
 package hu.dominikvaradi.sociallybackend.flows.security.service;
 
+import hu.dominikvaradi.sociallybackend.flows.common.config.ApplicationProperties;
 import hu.dominikvaradi.sociallybackend.flows.security.domain.RefreshToken;
+import hu.dominikvaradi.sociallybackend.flows.security.repository.RefreshTokenRepository;
 import hu.dominikvaradi.sociallybackend.flows.user.domain.User;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Instant;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface RefreshTokenService {
-	Optional<RefreshToken> findRefreshTokenByToken(UUID token);
+@RequiredArgsConstructor
+@Transactional
+@Service
+public class RefreshTokenService {
+	private final RefreshTokenRepository refreshTokenRepository;
+	private final ApplicationProperties applicationProperties;
 
-	Optional<RefreshToken> findRefreshTokenByUser(User user);
+	public Optional<RefreshToken> findRefreshTokenByToken(UUID token) {
+		return refreshTokenRepository.findByToken(token);
+	}
 
-	RefreshToken createRefreshToken(User user);
+	public RefreshToken createRefreshToken(User user) {
+		RefreshToken createdRefreshToken = RefreshToken.builder()
+				.user(user)
+				.expiresAt(Instant.now().plusMillis(applicationProperties.getJwt().getRefreshTokenExpirationInMillis()))
+				.build();
 
-	boolean isRefreshTokenExpired(RefreshToken refreshToken);
+		return refreshTokenRepository.save(createdRefreshToken);
+	}
 
-	void deleteRefreshToken(RefreshToken refreshToken);
+	public RefreshToken createAndDeleteExistingRefreshToken(User user) {
+		Optional<RefreshToken> existingRefreshToken = refreshTokenRepository.findByUser(user);
+		if (existingRefreshToken.isPresent()) {
+			refreshTokenRepository.delete(existingRefreshToken.get());
+			refreshTokenRepository.flush();
+		}
+
+		RefreshToken createdRefreshToken = RefreshToken.builder()
+				.user(user)
+				.expiresAt(Instant.now().plusMillis(applicationProperties.getJwt().getRefreshTokenExpirationInMillis()))
+				.build();
+
+		return refreshTokenRepository.save(createdRefreshToken);
+	}
+
+	public boolean isRefreshTokenExpired(RefreshToken refreshToken) {
+		return refreshToken.getExpiresAt().compareTo(Instant.now()) < 0;
+	}
+
+	public void deleteRefreshToken(RefreshToken refreshToken) {
+		refreshTokenRepository.delete(refreshToken);
+	}
 }
